Divorce handleError form the store.
The definition of handleError as an action itself is a leftover from
earlier incarnations of the code - at this point, it merely a function
that does some initial processing of the error being input and the set
of handlers that shuld be applide to the error. The error handlers are
wrapped so they are dispatch()ed on the store.

As such, it is ready to become a simple funciton. Make it so.